+++ cmsplugin_blocks/factories/__init__.py
@@ -1,6 +1,10 @@
+from cmsplugin_blocks.factories.card import CardFactory
 from cmsplugin_blocks.factories.hero import HeroFactory
+from cmsplugin_blocks.factories.slider import SliderFactory
 
 
 __all__ = [
+    'CardFactory',
     'HeroFactory',
+    'SliderFactory',
 ]