+++ HISTORY.rst
@@ -2,6 +2,13 @@
 Changelog
 =========
 
+Version 0.6.0 - Unreleased
+--------------------------
+
+* Added test coverage and factories for every plugins;
+* Modified every plugin templates to be cleaner and flawless;
+* Added ``title`` attribute to SlideItem model;
+
 Version 0.5.2 - 2019/05/18
 --------------------------